feat: [session-persistence-dashboard-checkpoint-guard-001] - [Gate restore on session sync preflight]
